+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.7] - 2021-08-16
+### Added
+### Changed
+- Se modifico el Historial Académico
+- Se modifico el Centralizador Calificaciones
+### Removed
+### Fixed
+
 ## [1.2.6] - 2021-08-05
 ### Added
 ### Changed